requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.4
decorator==5.2.1
Deprecated==1.2.18
et_xmlfile==2.0.0
fabric==3.2.2
fake==0.8
Faker==37.4.0
iniconfig==2.1.0
invoke==2.2.0
mimesis
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
paramiko==3.5.1
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
PyPDF2==3.0.1
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
wrapt==1.17.2
python-dotenv==1.1.1